+++ app/src/main/java/com/krisna/storycircle/presentation/activity/auth/RegisterActivity.kt
@@ -1,20 +1,15 @@
 package com.krisna.storycircle.presentation.activity.auth
 
 import android.os.Bundle
-import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
 import com.krisna.storycircle.databinding.ActivityRegisterBinding
-import com.krisna.storycircle.di.ViewModelFactory
 import com.krisna.storycircle.presentation.viewmodel.AuthViewModel
+import org.koin.androidx.viewmodel.ext.android.viewModel
 
 class RegisterActivity : AppCompatActivity() {
 
     private lateinit var binding: ActivityRegisterBinding
-    private val authViewModel: AuthViewModel by viewModels(
-        factoryProducer = {
-            ViewModelFactory.getInstance(this)
-        }
-    )
+    private val authViewModel: AuthViewModel by viewModel()
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
