+++ app/src/main/java/com/krisna/storycircle/presentation/viewmodel/AuthViewModel.kt
@@ -1,12 +1,10 @@
 package com.krisna.storycircle.presentation.viewmodel
 
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.viewModelScope
+import androidx.lifecycle.*
 import com.krisna.storycircle.data.model.response.RegisterResponse
 import com.krisna.storycircle.data.repository.StoryCircleRepository
 import kotlinx.coroutines.launch
+import org.koin.mp.KoinPlatform.getKoin
 
 class AuthViewModel(
     private val storyCircleRepository: StoryCircleRepository
@@ -33,6 +31,12 @@ class AuthViewModel(
                 _isLoading.value = false
                 _errorMessage.value = it.message
             }
+        }
+    }
+
+    companion object {
+        fun createViewModel(viewModelStoreOwner: ViewModelStoreOwner): AuthViewModel {
+            return ViewModelProvider(viewModelStoreOwner, getKoin().get())[AuthViewModel::class.java]
         }
     }
 }
